requirements.txt
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
python-dateutil==2.9.0
python-dotenv==1.0.1